aula_03.4

+++ Aula_3/Atividade_03/questao7/weg3.js
@@ -12,9 +12,9 @@ const conexao = mysql.createConnection({
 function cadastrarFuncionarios() {
 
     const nome = readline.question("Digite o nome do funcionario: ");
-    const cargo = readline.question("Digite o telefone do funcionarios: ");
+    const cargo = readline.question("Digite o cargo do funcionarios: ");
 
-    const insert = "INSERT INTO funcionarios (nome, telefone) VELUES (?,?)";
+    const insert = "INSERT INTO funcionarios (nome,cargo) VALUES (?,?)";
 
     conexao.query(insert, [nome, cargo], function (erro) {
 
@@ -34,40 +34,47 @@ function excluirFuncionarios() {
 
     const id = readline.question("Digite o id do funcionarios: ");
 
-    const deletar = "DELETE FROM funcionarios WHERE id = ?";
+    const confirmar = readline.question("Deseja realmente excluir este funcionário? (S/N):");
 
-    conexao.query(deletar, [id], function (erro,resultado) {
+    if (confirmar.toUpperCase()==="S") {
 
-        if (erro) {
-            console.log("Erro ao excluir o funcionarios.");
-            console.log(erro);
-        } else if (resultado.affectedRows ===0) {
-            console.log("funcionarios não encontrado.");
-        } else {
-            console.log("funcionarios excluido com sucesso!!!");
-        }
+        const deletar = "DELETE FROM funcionarios WHERE id = ?";
 
-        menu();
-    });
+        conexao.query(deletar, [id], function (erro,resultado) {
+
+            if (erro) {
+                console.log("Erro ao excluir o funcionarios.");
+                console.log(erro);
+            } else if (resultado.affectedRows ===0) {
+                console.log("Funcionarios não encontrado.");
+            } else {
+                console.log("Funcionarios excluido com sucesso!!!");
+            }
+
+        });
+    } else {
+        console.log("Funcionario não excluido!")
+    }
+    menu();
 }
 
 
-function listarCliente() {
+function listarFuncionarios() {
 
     const sql = "SELECT * FROM funcionarios";
 
-    conexao.query(sql, function (erro, clientes) {
+    conexao.query(sql, function (erro, funcionarios) {
 
         if (erro) {
             console.log("Erro ao listar os funcionarios.");
             console.log(erro);
         } else {
-            console("\n===== CLIENTES =====");
+            console.log("\n===== FUNCIONARIOS =====");
             funcionarios.forEach(function (funcionarios) {
                 console.log(
                     funcionarios.id +" - "+
                     funcionarios.nome +" - "+
-                    funcionarios.telefone
+                    funcionarios.cargo
                 );
             });
         }
@@ -83,22 +90,22 @@ function menu() {
     console.log("\n===== MENU =====");
     console.log("1 - Cadastrar funcionarios");
     console.log("2 - Excluir funcionarios");
-    console.log("funcionarios");
+    console.log("3 - Listar funcionarios");
     console.log("0 - Sair");
 
     const opcao = readline.questionInt("Escolha uma opção: ");
 
     if (opcao === 1){
 
-        cadastrarCliente();
+        cadastrarFuncionarios();
 
     } else if (opcao === 2) {
 
-        excluirCliente();
+        excluirFuncionarios();
 
     } else if (opcao === 3) {
 
-        listarCliente();
+        listarFuncionarios();
 
     } else if (opcao === 0) {
